refactor(query): rename _client.ts to client.ts
The underscore prefix was unnecessary ceremony. The separation between
client.ts (infrastructure) and index.ts (public API) already communicates
intent clearly without the visual noise.
