+++ apps/whispering/src/lib/query/desktop/tray.ts
@@ -2,7 +2,7 @@ import { Ok, type Result } from 'wellcrafted/result';
 import type { WhisperingRecordingState } from '$lib/constants/audio';
 import { WhisperingErr, type WhisperingError } from '$lib/result';
 import { desktopServices } from '$lib/services';
-import { defineMutation } from '../_client';
+import { defineMutation } from '../client';
 
 const setTrayIconKeys = {
 	setTrayIcon: ['setTrayIcon', 'setTrayIcon'] as const,
