+++ apps/whispering/src/lib/query/isomorphic/download.ts
@@ -3,7 +3,7 @@ import { WhisperingErr, type WhisperingError } from '$lib/result';
 import { services } from '$lib/services';
 import type { Recording } from '$lib/services/isomorphic/db';
 import type { DownloadServiceError } from '$lib/services/isomorphic/download';
-import { defineMutation } from '../_client';
+import { defineMutation } from '../client';
 
 export const download = {
 	downloadRecording: defineMutation({
