+++ apps/whispering/src/lib/query/isomorphic/notify.ts
@@ -3,7 +3,7 @@ import { dev } from '$app/environment';
 import { notificationLog } from '$lib/components/NotificationLog.svelte';
 import { services } from '$lib/services';
 import type { UnifiedNotificationOptions } from '$lib/services/isomorphic/notifications/types';
-import { defineMutation } from '../_client';
+import { defineMutation } from '../client';
 
 // Create a mutation for a specific variant
 const createNotifyMutation = (
